WIP - added basic spec for operation and moved result class to its own file
