+++ services/comm/app/operations/event_process.rb
@@ -3,12 +3,12 @@
 # NOTE: Might be more worthy to implement this using TRB operation instead
 class EventProcess < Trailblazer::Activity::FastTrack
   # rubocop:disable Style/SignalException
-  step :find_event, input: [:params], output: %i[event template campaign]
+  step :find_event
   fail :event_not_found, fail_fast: true
   step :process_event
   # rubocop:enable Style/SignalException
 
-  def find_event(ctx, params)
+  def find_event(ctx, params:)
     event = ::Event.find_by(params)
     return false unless event
 
@@ -17,11 +17,12 @@ class EventProcess < Trailblazer::Activity::FastTrack
     ctx[:campaign] = event.campaign
   end
 
-  def event_not_found(_ctx)
+  def event_not_found(ctx, params:)
+    ctx[:errors] ||= []
     ctx[:errors] << "{EventProcess} Can't find event (#{params}) for tenant"
   end
 
-  def process_event(_ctx, event:, template:, campaign:)
+  def process_event(_ctx, event:, template:, campaign:, **)
     event.process!
     event.users.each do |user|
       content = template.render(user, campaign)
